feat : 달력 화면의 component 분리 및 graph 화면과 연도/월 연동 #41

+++ app/src/main/java/com/seom/accountbook/data/local/AccountDao.kt
@@ -199,7 +199,7 @@ class AccountDao(
                 accounts.add(
                     CalendarEntity(
                         date = cursor.getInt(0),
-                        count = cursor.getInt(1),
+                        count = cursor.getLong(1),
                         type = cursor.getInt(2)
                     )
                 )
